pyproject.toml
[project]
name = "web-to-markdown"
version = "0.1.0"
description = "A tool to convert web pages to markdown using uilixee/hero"
dependencies = [
    "playwright",
    "python-dotenv",
    "html2text",
    "asyncio",
]
requires-python = ">=3.13"

[tool.ruff]
target-version = "py313"
line-length = 88 